% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nlx_synch.R
\name{get_nlx_synch}
\alias{get_nlx_synch}
\title{Synchronous NLX API Wrapper}
\usage{
get_nlx_synch(state_or_territory, page = 1)
}
\arguments{
\item{state_or_territory}{two digit state or territory code. This parameter is not case sensitive.}

\item{page}{the page number of results to return.}
}
\value{
a tibble
}
\description{
`get_nlx_synch()` accesses the synchronous NLX REST API for paging through data 50 records at a time.
}
\examples{
get_nlx_synch(state_or_territory = 'ky', page = 1)


}
