% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nlx.R
\name{nlx_download_report}
\alias{nlx_download_report}
\title{Downloads Data From Requested Report URL}
\usage{
nlx_download_report(report_status_url, sleep_time = 2, silently = TRUE)
}
\arguments{
\item{report_status_url}{URL of the requested report as a string.}

\item{sleep_time}{The amount of time in seconds between queries to the NLX
API. Avoids rate limitations. Defaults to two seconds.}

\item{silently}{Provides updates on status on the progress of the report
between queries if true. Defaults to `FALSE`}
}
\value{
a tibble
}
\description{
`nlx_download_report()` checks the status of a requested report, and when complete, downloads the requested data into a tibble
}
\examples{
df <- nlx_download_report('')
}
