% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nlx.R
\name{nlx_report_request}
\alias{nlx_report_request}
\title{Request Report From NLX Asynchronous API}
\usage{
nlx_report_request(
  state_or_territory,
  start_date,
  end_date = NULL,
  date_column = "created_date"
)
}
\arguments{
\item{state_or_territory}{Two digit state or territory code. This parameter
is not case sensitive.}

\item{start_date}{The start date for the creation_date of job postings
returned by the query (inclusive).}

\item{end_date}{The end date for the creation_date of job postings returned
by the query (inclusive). If no end_date is included the query defaults to
one full day beginning on the start_date.Please note that you cannot
request more than 35 days of data at a time Dates are midnight to midnight,
so 2021-06-01 to 2021-06-02 is one day of data, for 2021-06-01}

\item{date_column}{The date field that start_date and end_date filter by
Options are created_date, date_compiled, last_updated_date, or
date_acquired. Defaults to created_date.}
}
\value{
string containing the url of the requested report
}
\description{
`nlx_report_request()` Requests a report from the NLX asynchronous API and
returns the url of the requested report.
}
\examples{
request_url <- nlx_report_request(state_or_territory='OH',
start_date = '2021-06-05', end_date = '2021-06-10')

}
